Make method extensions
